Migrate storage tests to storage
